Refactor app.py into blueprints and modules
